refactor(content): extract CollectionsSidebar into its own file
content-browser.tsx is a 2700+ line God component; CollectionsSidebar
and CollectionRow are pure-props (no shared closures/local state) and
form a self-contained sidebar nav unit, so they move out unchanged.
